+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - feat: `diff` command, which displays changes between version (or the current HEAD) as pretty diff output
 - feat: add `GPM_AI_CHAT_TEMPERATURE` environment variable, which defines the custom temperature value for AI chat (operations)
 - feat: add `--temperature` flag to `chat` command, which can define the initial temperature value for the command
+- feat: `setup updater` command, which installs a shell script called `gpm-update` in `$HOME/.gpm/bin` folder of a UNIX environment, like *BSD, Linux or MacOS
 - refactor: improve prompting in `chat` command
 - refactor: `pack` command now outputs progress with pretty bars